RQ 1 Update
Changed DataFrame configuration for ASEAN where it does not include the Philippines. Furthermore, the ttest method was also changed into ttest_ind() instead of ttest_rel() to signify that the data of the Philippines and the ASEAN neighbors are unpaired therefore effectively indendent of each other.

Further improvements on writing of explanations are to be made on the upcoming versions.